requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.3
pydantic[email]==2.8.2      
pydantic-settings==2.4.0
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
supabase==2.3.4
python-multipart==0.0.9
python-dotenv==1.0.1
